+++ packages/types-bundle-config/eslint.config.js
@@ -1,1 +1,1 @@
-module.exports = require("@rnx-kit/eslint-config");
+module.exports = require("@rnx-kit/eslint-config/types-only");