+++ packages/types-bundle-config/src/allPlatforms.ts
@@ -3,6 +3,7 @@
  */
 
 // Possible values for AllPlatforms
+// eslint-disable-next-line @rnx-kit/type-definitions-only
 export const ALL_PLATFORM_VALUES = [
   "android",
   "ios",